Mypage (#63)

+++ src/pages/MyPage/MyPage.tsx
@@ -1,21 +1,30 @@
-import React, { FC, useCallback } from 'react';
+import React, { FC, useCallback, useState, useRef, ChangeEvent } from 'react';
 import { useNavigate } from 'react-router';
-import { useMutation, useQuery } from 'react-query';
 
-import { Container, UpdateBtn, UserInfo, LogoutBtn, User, Locations, LocationCard } from './styles';
+import { useMutation, useQuery } from 'react-query';
+import { getMeHelper } from 'src/query';
+import { createUserImg } from 'src/api/auth';
+import { Container, UpdateBtn, UserInfo, LogoutBtn, User, Locations, LocationCard, InputImg } from './styles';
 import HeaderWithLeftArrow from 'src/components/HeaderWithLeftArrow';
 import { Path } from 'src/Constants';
-import { useMeState, useMyLocationState } from 'src/atoms';
+import { useMyLocationState } from 'src/atoms';
 import { getMyLocation, deleteMyLocation } from 'src/api/marker';
+import NickNameUpdateModal from 'src/components/NicknameModal';
 
 import mypage from 'src/assets/mypage/user-img.png';
+import camera from 'src/assets/mypage/camera.png';
 
 const MyPage: FC = () => {
-    const { me, setMe } = useMeState();
+    const { data: me } = getMeHelper.useQuery();
     const navigate = useNavigate();
-
-    const { data: locations } = useQuery('getLocations', () => getMyLocation({ offset: 0, limit: 50 }));
     const { setLocations } = useMyLocationState();
+    const [nicknamePopup, setNicknamePopup] = useState(false);
+
+    const { data: locations } = useQuery('getLocations', () => getMyLocation({ offset: 0, limit: 50 }), {
+        onSuccess: response => {
+            setLocations(response);
+        }
+    });
     const { mutate: deleteLocation } = useMutation(deleteMyLocation, {
         onMutate: ({ addressId }) => {
             setLocations(locations => {
@@ -36,8 +45,13 @@ const MyPage: FC = () => {
     );
 
     const onDeleteClick = useCallback(
-        (addressId: number) => {
-            deleteLocation({ addressId });
+        async (addressId: number) => {
+            const confirm = window.confirm('해당 장소를 삭제하시겠습니까?');
+
+            if (confirm) {
+                await deleteLocation({ addressId });
+                window.alert('삭제되었습니다.');
+            }
         },
         [deleteLocation]
     );
@@ -48,9 +62,33 @@ const MyPage: FC = () => {
 
     const logout = useCallback(() => {
         localStorage.removeItem('token');
-        setMe(undefined);
+        getMeHelper.setQueryData(undefined);
         navigate(Path.login);
-    }, [navigate, setMe]);
+    }, [navigate]);
+
+    const onChange = useCallback(async (e: ChangeEvent<HTMLInputElement>) => {
+        const img = e.target.files?.[0];
+        if (!img) return;
+        const formData = new FormData();
+        formData.append('file', img);
+
+        const resultThumnail = await createUserImg({ file: formData });
+
+        getMeHelper.setQueryData(me => {
+            if (!me) return;
+            return { ...me, thumbnail: resultThumnail };
+        });
+    }, []);
+
+    const updateNickname = useCallback(() => {
+        setNicknamePopup(true);
+    }, [setNicknamePopup]);
+
+    const inputRef = useRef<HTMLInputElement>(null);
+
+    const onClickFileInput = useCallback(() => {
+        inputRef.current?.click();
+    }, []);
 
     return (
         <Container>
@@ -61,9 +99,13 @@ const MyPage: FC = () => {
             </HeaderWithLeftArrow>
             <UserInfo>
                 <User>
-                    <img className='mypage-img' src={me?.thumbnail || mypage} />
+                    <input ref={inputRef} accept='image/jpg,impge/png,image/jpeg,image/gif' className='profile_img' type='file' onChange={onChange} />
+                    <InputImg>
+                        <img className='mypage-img' src={me?.thumbnail || mypage} onClick={() => onClickFileInput()} />
+                        <img className='upload-img' src={camera} onClick={() => onClickFileInput()} />
+                    </InputImg>
                     <div className='user-txt'>{me?.nickname}</div>
-                    <UpdateBtn>수정</UpdateBtn>
+                    <UpdateBtn onClick={updateNickname}>수정</UpdateBtn>
                 </User>
 
                 <LogoutBtn onClick={logout}>로그아웃</LogoutBtn>
@@ -72,7 +114,7 @@ const MyPage: FC = () => {
                 <h2>저장한 장소</h2>
             </div>
             <Locations>
-                <div style={{ overflowY: 'scroll' }}>
+                <div>
                     {locations?.map(({ id, name, address, roadAddress, addressId }) => (
                         <LocationCard key={id}>
                             <div style={{ display: 'flex', flexDirection: 'column' }} onClick={() => onClickLocation(addressId)}>
@@ -88,6 +130,7 @@ const MyPage: FC = () => {
                     ))}
                 </div>
             </Locations>
+            {nicknamePopup && <NickNameUpdateModal setClose={() => setNicknamePopup(false)} />}
         </Container>
     );
 };
